// client/src/types/users.ts
export interface userResponse {
  isAdmin: boolean;
  _id: string;
  username: string;
  email: string;
  password: string;
  __v: number;
  createdAt?: Date;
  updatedAt?: Date;
}
export interface UserType {
  isAdmin: boolean;
  _id: string;
  username: string;
  email: string;
  password: string;
  __v: number;
  createdAt?: Date;
  updatedAt?: Date;
}